Add files via upload
This is the code made for the ESP8266. It receives the string passed by the arduino board and sent it to an internal webpage managed by littleFS and also through the influxdb making a local database and allowing displaying data using Graphana. 

The code allows sending on the local server different events# that are used to manage the data and to show them in charts.