FRESCO_I/O_desc
A short description about the Input and Output Pin available on the FRESCO board. It provides some specification about the pin if they are digital or analog, how they are named inside the firmware and their name according to the file saved into the SD card or to be recalled for the PID set point value.