Update of Webapp guide
The new guide includes some information about the use of settings in the web app and how to set some entries such as: NTP, TZ-INFO, B value, Send data transfer time, displaying time, saving time. If the PCool shield is recognized it allows also to set the desired reference for the PID and the sample surface to correctly evaluate the net cooling power density.